refactor: 💡 move exceptions to libapi
Also: ApiCustomError is now ApiError. And UnexpectedError, which was a
collision with the CacheableResponse, is now UnexpectedApiError
