+++ services/api/src/api/routes/valid.py
@@ -5,6 +5,7 @@ import logging
 from dataclasses import dataclass, field
 from typing import List, Set, TypedDict
 
+from libapi.exceptions import UnexpectedApiError
 from libapi.utils import Endpoint
 from libcommon.processing_graph import ProcessingGraph, ProcessingStep
 from libcommon.prometheus import StepProfiler
@@ -12,7 +13,7 @@ from libcommon.simple_cache import get_valid_datasets
 from starlette.requests import Request
 from starlette.responses import Response
 
-from api.utils import UnexpectedError, get_json_api_error_response, get_json_ok_response
+from api.utils import get_json_api_error_response, get_json_ok_response
 
 
 class ValidContent(TypedDict):
@@ -72,6 +73,8 @@ def create_valid_endpoint(
                     return get_json_ok_response(content, max_age=max_age_long)
             except Exception as e:
                 with StepProfiler(method="valid_endpoint", step="generate API error response"):
-                    return get_json_api_error_response(UnexpectedError("Unexpected error.", e), max_age=max_age_short)
+                    return get_json_api_error_response(
+                        UnexpectedApiError("Unexpected error.", e), max_age=max_age_short
+                    )
 
     return valid_endpoint
