+++ services/api/src/api/utils.py
@@ -1,75 +1,16 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright 2022 The HuggingFace Authors.
 
-import logging
 from http import HTTPStatus
-from typing import Any, List, Literal, Optional
+from typing import Any, List, Optional
 
+from libapi.exceptions import ResponseNotFoundError, ResponseNotReadyError
 from libcommon.dataset import get_dataset_git_revision
 from libcommon.exceptions import CustomError
 from libcommon.orchestrator import DatasetOrchestrator
 from libcommon.processing_graph import ProcessingGraph, ProcessingStep
 from libcommon.utils import Priority, orjson_dumps
 from starlette.responses import JSONResponse, Response
-
-ApiErrorCode = Literal[
-    "InvalidParameter",
-    "MissingRequiredParameter",
-    "MissingProcessingStepsError",
-    "ResponseNotFound",
-    "ResponseNotReady",
-    "UnexpectedError",  # also in libcommon.exceptions
-]
-
-
-class ApiCustomError(CustomError):
-    """Base class for exceptions in this module."""
-
-    def __init__(
-        self,
-        message: str,
-        status_code: HTTPStatus,
-        code: ApiErrorCode,
-        cause: Optional[BaseException] = None,
-        disclose_cause: bool = False,
-    ):
-        super().__init__(message, status_code, str(code), cause, disclose_cause)
-
-
-class MissingRequiredParameterError(ApiCustomError):
-    """Raised when a required parameter is missing."""
-
-    def __init__(self, message: str):
-        super().__init__(message, HTTPStatus.UNPROCESSABLE_ENTITY, "MissingRequiredParameter")
-
-
-class InvalidParameterError(ApiCustomError):
-    """Raised when a parameter has an invalid value."""
-
-    def __init__(self, message: str):
-        super().__init__(message, HTTPStatus.UNPROCESSABLE_ENTITY, "InvalidParameter")
-
-
-class ResponseNotReadyError(ApiCustomError):
-    """Raised when the response has not been processed yet."""
-
-    def __init__(self, message: str):
-        super().__init__(message, HTTPStatus.INTERNAL_SERVER_ERROR, "ResponseNotReady")
-
-
-class ResponseNotFoundError(ApiCustomError):
-    """Raised when the response has not been found."""
-
-    def __init__(self, message: str):
-        super().__init__(message, HTTPStatus.NOT_FOUND, "ResponseNotFound")
-
-
-class UnexpectedError(ApiCustomError):
-    """Raised when the server raised an unexpected error."""
-
-    def __init__(self, message: str, cause: Optional[BaseException] = None):
-        super().__init__(message, HTTPStatus.INTERNAL_SERVER_ERROR, "UnexpectedError", cause)
-        logging.error(message, exc_info=cause)
 
 
 class OrjsonResponse(JSONResponse):
