+++ services/api/tests/routes/test_rows.py
@@ -1,3 +1,6 @@
+# SPDX-License-Identifier: Apache-2.0
+# Copyright 2023 The HuggingFace Authors.
+
 import os
 import shutil
 import time
